+++ nanomesh/mesh/__init__.py
@@ -1,12 +1,13 @@
-from ._base import GenericMesh
 from ._line import LineMesh
+from ._mesh import Mesh
+from ._mixin import PruneZ0Mixin
 from ._tetra import TetraMesh
 from ._triangle import TriangleMesh
 
 __all__ = [
     'LineMesh',
-    'GenericMesh',
-    'registry',
+    'Mesh',
     'TetraMesh',
     'TriangleMesh',
+    'PruneZ0Mixin',
 ]
